Change access token to session storage
As it's a little exposed to script injection (if that occurs),
so best not keep it longer than absolutely required.

Full solution would be to run a proxy server and store in httponly
cookie. It would require proxying all api requests that required auth.
I don't have a server to host that, and it seems a little overkill for a
prototype.